fixed mismatch int - long

+++ src/main/java/pt/ipleiria/estg/dei/ei/dae/project/ejbs/ClientBean.java
@@ -27,7 +27,7 @@ public class ClientBean {
         return (List<Client>) entityManager.createNamedQuery("getAllClients").getResultList();
     }
 
-    public Client findClient(long id) {
+    public Client findClient(int id) {
         return entityManager.find(Client.class, id);
     }
 }
